修复头文件中的 clang-tidy 警告（全量零警告）
- C 数组改 std::array：Fmt 格式缓冲、FixedBuffer 数据区、
  LogStream 数字字符表、Logger tid 字符串、LogFile setbuffer 缓冲、
  HttpConnection iovec 数组
- 枚举指定 std::uint8_t 底层类型：ParseResult/ParseState/Method/Version/LogLevel
- HttpRequest 成员改类内默认初始化（移除冗余构造列表项）
- 整型乘法显式宽化到 size_t：ThreadPool 线程数、Buffer/FixedBuffer/
  LogFile 缓冲大小、请求限制默认值
- ThreadPool::submit 用 lambda 捕获替代 std::bind
- HeaderFilterRegex 修正为匹配绝对路径，头文件检查对 cpp 编译单元生效

验证：clang-tidy 全量（cpp+hpp）零警告；Debug/ASan 构建零警告；
96 单元 + 32 集成测试全部通过。

+++ src/net/Buffer.hpp
@@ -45,7 +45,7 @@ class Buffer {
 
 public:
     Buffer()
-        : Buffer(64 * 1024) {}
+        : Buffer(static_cast<size_t>(64) * 1024) {}
 
     explicit Buffer(const size_t initSize)
         : buffer_(initSize) {}
